KediaHomes/requirements.txt
asgiref==3.4.1
attrs==19.3.0
Automat==0.8.0
awscli==1.18.69
blinker==1.4
boto3==1.18.29
botocore==1.21.29
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
dbus-python==1.2.16
distro==1.4.0
Django==3.2.6
django-ckeditor==6.1.0
django-cors-headers==3.8.0
django-js-asset==1.2.2
django-storages==1.11.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
docutils==0.15.2
entrypoints==0.3
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
Jinja2==2.10.1
jmespath==0.10.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==1.1.0
more-itertools==4.2.0
mysqlclient==2.0.3
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
pexpect==4.6.0
Pillow==8.3.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
Pygments==2.3.1
PyHamcrest==1.9.0
PyJWT==2.1.0
pymacaroons==0.13.0
PyMySQL==1.0.2
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyserial==3.4
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.3.1
razorpay==1.2.0
requests==2.22.0
requests-unixsocket==0.2.0
roman==2.0.0
rsa==4.0
s3transfer==0.5.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sqlparse==0.4.1
ssh-import-id==5.10
Twisted==18.9.0
urllib3==1.25.8
wadllib==1.3.3
zipp==1.0.0
zope.interface==4.7.1
